+++ php-1.2.1/submitDependencies.php
@@ -24,13 +24,13 @@
 <script src="/externalDependencies/bootbox/bootbox.locales.min.js"></script>
 
 <!--my scripts-->
-<script src="js-1.2.0/errHandler.js"></script>
-<script src="js-1.2.0/settings.js"></script>
-<script src="js-1.2.0/submitComponents/imageStore.js"></script>
-<script src="js-1.2.0/submitComponents/inputs.js"></script>
-<script src="js-1.2.0/submitComponents/calcVars.js"></script>
-<script src="js-1.2.0/submitComponents/docDef.js"></script>
-<script src="js-1.2.0/submit.js"></script>
+<script src="js-1.2.1/errHandler.js"></script>
+<script src="js-1.2.1/settings.js"></script>
+<script src="js-1.2.1/submitComponents/imageStore.js"></script>
+<script src="js-1.2.1/submitComponents/inputs.js"></script>
+<script src="js-1.2.1/submitComponents/calcVars.js"></script>
+<script src="js-1.2.1/submitComponents/docDef.js"></script>
+<script src="js-1.2.1/submit.js"></script>
 
 <!--popovers and tooltips-->
 <script>